enable gci in golangci-lint
This will ensure that the imports are formatted with a consistent style.

goimports simply fixes the imports, but not the order.

gci is a tool that formats the imports in a consistent order,
but by design it does not fix the imports.

So we need to use both tools together.